refactor developer_portal into smaller modules
