+++ src-tauri/src/developer_portal/mod.rs
@@ -1,0 +1,8 @@
+pub mod hugo;
+pub mod repo;
+pub mod authors;
+
+// Re-exporting commands for easy access
+pub use hugo::*;
+pub use repo::*;
+pub use authors::*;
